change /success link to /manage
