+++ apps/web/lib/hooks/usePublicPage.ts
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 
 export default function usePublicPage() {
   const router = useRouter();
-  const isPublicPage = ["/[user]", "/success", "/cancel", "/reschedule"].find((route) =>
+  const isPublicPage = ["/[user]", "/manage", "/cancel", "/reschedule"].find((route) =>
     router.pathname.startsWith(route)
   );
   return isPublicPage;
