+++ apps/web/playwright/dynamic-booking-pages.e2e.ts
@@ -37,7 +37,7 @@ test("dynamic booking", async ({ page, users }) => {
     await page.locator('[data-testid="confirm-reschedule-button"]').click();
     await page.waitForNavigation({
       url(url) {
-        return url.pathname === "/success";
+        return url.pathname === "/manage";
       },
     });
     await expect(page.locator("[data-testid=success-page]")).toBeVisible();
@@ -48,7 +48,7 @@ test("dynamic booking", async ({ page, users }) => {
     await page.locator('[data-testid="cancel"]').first().click();
     await page.waitForNavigation({
       url: (url) => {
-        return url.pathname.startsWith("/success");
+        return url.pathname.startsWith("/manage");
       },
     });
     // --- fill form
@@ -56,7 +56,7 @@ test("dynamic booking", async ({ page, users }) => {
 
     await page.waitForNavigation({
       url: (url) => {
-        return url.pathname.startsWith("/success");
+        return url.pathname.startsWith("/manage");
       },
     });
     const successHeadling = await page.locator('[data-testid="success-headline"]').innerText();
